Use component name instead of id in api definitions
