+++ golem-worker-service-base/src/gateway_api_definition/http/api_definition_request.rs
@@ -16,7 +16,8 @@ use crate::gateway_api_definition::http::{AllPathPatterns, MethodPattern, Route}
 use crate::gateway_api_definition::{ApiDefinitionId, ApiVersion};
 use crate::gateway_binding::GatewayBinding;
 use crate::gateway_middleware::HttpCors;
-use crate::gateway_security::SecuritySchemeReference;
+use crate::gateway_security::{SecuritySchemeIdentifier, SecuritySchemeReference};
+use golem_api_grpc::proto::golem::apidefinition as grpc_apidefinition;
 
 // HttpApiDefinitionRequest corresponds to the user facing http api definition.
 // It has security at the global level, which is following OpenAPI style of defining security at the root level.
@@ -28,6 +29,47 @@ pub struct HttpApiDefinitionRequest {
     pub version: ApiVersion,
     pub routes: Vec<RouteRequest>,
     pub draft: bool,
+}
+
+impl TryFrom<grpc_apidefinition::v1::ApiDefinitionRequest> for HttpApiDefinitionRequest {
+    type Error = String;
+
+    fn try_from(value: grpc_apidefinition::v1::ApiDefinitionRequest) -> Result<Self, Self::Error> {
+        let mut global_securities = vec![];
+        let mut route_requests = vec![];
+
+        match value.definition.ok_or("definition is missing")? {
+            grpc_apidefinition::v1::api_definition_request::Definition::Http(http) => {
+                for route in http.routes {
+                    let route_request =
+                        crate::gateway_api_definition::http::RouteRequest::try_from(route)?;
+                    if let Some(security) = &route_request.security {
+                        global_securities.push(security.clone());
+                    }
+
+                    route_requests.push(route_request);
+                }
+            }
+        };
+
+        let id = value.id.ok_or("Api Definition ID is missing")?;
+
+        let security = if global_securities.is_empty() {
+            None
+        } else {
+            Some(global_securities)
+        };
+
+        let result = Self {
+            id: ApiDefinitionId(id.value),
+            version: ApiVersion(value.version),
+            routes: route_requests,
+            draft: value.draft,
+            security,
+        };
+
+        Ok(result)
+    }
 }
 
 // In a RouteRequest, security is defined at the outer level
@@ -61,5 +103,38 @@ impl From<Route> for RouteRequest {
                 .map(|x| SecuritySchemeReference::from(x.security_scheme_with_metadata)),
             cors: cors_middleware,
         }
+    }
+}
+
+impl TryFrom<grpc_apidefinition::HttpRoute> for RouteRequest {
+    type Error = String;
+
+    fn try_from(value: grpc_apidefinition::HttpRoute) -> Result<Self, Self::Error> {
+        let path = AllPathPatterns::parse(value.path.as_str()).map_err(|e| e.to_string())?;
+        let binding = value.binding.ok_or("binding is missing")?;
+        let method: MethodPattern = value.method.try_into()?;
+
+        let gateway_binding = GatewayBinding::try_from(binding)?;
+        let security = value.middleware.clone().and_then(|x| x.http_authentication);
+
+        let security = security.and_then(|x| {
+            x.security_scheme.map(|x| SecuritySchemeReference {
+                security_scheme_identifier: SecuritySchemeIdentifier::new(x.scheme_identifier),
+            })
+        });
+
+        let cors = value.middleware.and_then(|x| x.cors);
+
+        let cors = cors.map(HttpCors::try_from).transpose()?;
+
+        let result = Self {
+            method,
+            path,
+            binding: gateway_binding,
+            security,
+            cors,
+        };
+
+        Ok(result)
     }
 }
