+++ golem-worker-service-base/src/gateway_api_definition/http/oas_api_definition.rs
@@ -14,6 +14,7 @@
 
 use crate::gateway_api_definition::http::HttpApiDefinitionRequest;
 use crate::gateway_api_definition::{ApiDefinitionId, ApiVersion};
+use crate::service::gateway::BoxConversionContext;
 use internal::*;
 use openapiv3::OpenAPI;
 use poem_openapi::registry::{MetaSchema, MetaSchemaRef};
@@ -21,11 +22,17 @@ use poem_openapi::types::{ParseError, ParseFromJSON, ParseFromYAML, ParseResult}
 use serde_json::Value;
 use std::borrow::Cow;
 
-pub struct OpenApiHttpApiDefinitionRequest(pub OpenAPI);
+pub struct OpenApiHttpApiDefinition(pub OpenAPI);
 
-impl OpenApiHttpApiDefinitionRequest {
-    pub fn to_http_api_definition_request(&self) -> Result<HttpApiDefinitionRequest, String> {
+impl OpenApiHttpApiDefinition {
+    // Convert an open api spec with golem extensions into a golem http api definition.
+    // This may make api calls to resolve component names used in the definition
+    pub async fn to_http_api_definition_request(
+        &self,
+        conversion_context: &BoxConversionContext<'_>,
+    ) -> Result<HttpApiDefinitionRequest, String> {
         let open_api = &self.0;
+
         let api_definition_id = ApiDefinitionId(get_root_extension_str(
             open_api,
             GOLEM_API_DEFINITION_ID_EXTENSION,
@@ -38,7 +45,7 @@ impl OpenApiHttpApiDefinitionRequest {
 
         let security = get_global_security(open_api);
 
-        let routes = get_routes(&open_api.paths)?;
+        let routes = get_routes(&open_api.paths, conversion_context).await?;
 
         Ok(HttpApiDefinitionRequest {
             id: api_definition_id,
@@ -50,11 +57,11 @@ impl OpenApiHttpApiDefinitionRequest {
     }
 }
 
-impl ParseFromJSON for OpenApiHttpApiDefinitionRequest {
+impl ParseFromJSON for OpenApiHttpApiDefinition {
     fn parse_from_json(value: Option<serde_json::Value>) -> ParseResult<Self> {
         match value {
             Some(value) => match serde_json::from_value::<openapiv3::OpenAPI>(value) {
-                Ok(openapi) => Ok(OpenApiHttpApiDefinitionRequest(openapi)),
+                Ok(openapi) => Ok(OpenApiHttpApiDefinition(openapi)),
                 Err(e) => Err(ParseError::<Self>::custom(format!(
                     "Failed to parse OpenAPI: {}",
                     e
@@ -68,11 +75,11 @@ impl ParseFromJSON for OpenApiHttpApiDefinitionRequest {
     }
 }
 
-impl ParseFromYAML for OpenApiHttpApiDefinitionRequest {
+impl ParseFromYAML for OpenApiHttpApiDefinition {
     fn parse_from_yaml(value: Option<Value>) -> ParseResult<Self> {
         match value {
             Some(value) => match serde_json::from_value::<openapiv3::OpenAPI>(value) {
-                Ok(openapi) => Ok(OpenApiHttpApiDefinitionRequest(openapi)),
+                Ok(openapi) => Ok(OpenApiHttpApiDefinition(openapi)),
                 Err(e) => Err(ParseError::<Self>::custom(format!(
                     "Failed to parse OpenAPI: {}",
                     e
@@ -86,7 +93,7 @@ impl ParseFromYAML for OpenApiHttpApiDefinitionRequest {
     }
 }
 
-impl poem_openapi::types::Type for OpenApiHttpApiDefinitionRequest {
+impl poem_openapi::types::Type for OpenApiHttpApiDefinition {
     const IS_REQUIRED: bool = true;
 
     type RawValueType = Self;
@@ -117,8 +124,11 @@ impl poem_openapi::types::Type for OpenApiHttpApiDefinitionRequest {
 }
 
 mod internal {
+
     use crate::gateway_api_definition::http::{AllPathPatterns, MethodPattern, RouteRequest};
-    use golem_common::model::{ComponentId, GatewayBindingType};
+
+    use crate::service::gateway::BoxConversionContext;
+    use golem_common::model::GatewayBindingType;
     use openapiv3::{OpenAPI, Operation, Paths, ReferenceOr};
     use rib::Expr;
     use serde_json::Value;
@@ -129,17 +139,16 @@ mod internal {
     use crate::gateway_middleware::{CorsPreflightExpr, HttpCors};
     use crate::gateway_security::{SecuritySchemeIdentifier, SecuritySchemeReference};
     use golem_service_base::model::VersionedComponentId;
-    use uuid::Uuid;
 
-    pub(crate) const GOLEM_API_DEFINITION_ID_EXTENSION: &str = "x-golem-api-definition-id";
-    pub(crate) const GOLEM_API_DEFINITION_VERSION: &str = "x-golem-api-definition-version";
+    pub(super) const GOLEM_API_DEFINITION_ID_EXTENSION: &str = "x-golem-api-definition-id";
+    pub(super) const GOLEM_API_DEFINITION_VERSION: &str = "x-golem-api-definition-version";
 
     // Legacy extension for worker bridge
-    pub(crate) const GOLEM_WORKER_GATEWAY_EXTENSION_LEGACY: &str = "x-golem-worker-bridge";
+    pub(super) const GOLEM_WORKER_GATEWAY_EXTENSION_LEGACY: &str = "x-golem-worker-bridge";
 
-    pub(crate) const GOLEM_API_GATEWAY_BINDING: &str = "x-golem-api-gateway-binding";
+    pub(super) const GOLEM_API_GATEWAY_BINDING: &str = "x-golem-api-gateway-binding";
 
-    pub(crate) fn get_global_security(open_api: &OpenAPI) -> Option<Vec<SecuritySchemeReference>> {
+    pub(super) fn get_global_security(open_api: &OpenAPI) -> Option<Vec<SecuritySchemeReference>> {
         open_api.security.as_ref().and_then(|requirements| {
             let global_security: Vec<_> = requirements
                 .iter()
@@ -157,7 +166,7 @@ mod internal {
             }
         })
     }
-    pub(crate) fn get_root_extension_str(
+    pub(super) fn get_root_extension_str(
         open_api: &OpenAPI,
         key_name: &str,
     ) -> Result<String, String> {
@@ -168,7 +177,7 @@ mod internal {
             .map(|x| x.to_string())
     }
 
-    pub(crate) fn get_root_extension_value(open_api: &OpenAPI, key_name: &str) -> Option<Value> {
+    pub(super) fn get_root_extension_value(open_api: &OpenAPI, key_name: &str) -> Option<Value> {
         open_api
             .extensions
             .iter()
@@ -176,7 +185,10 @@ mod internal {
             .map(|(_, v)| v.clone())
     }
 
-    pub(crate) fn get_routes(paths: &Paths) -> Result<Vec<RouteRequest>, String> {
+    pub(super) async fn get_routes(
+        paths: &Paths,
+        ctx: &BoxConversionContext<'_>,
+    ) -> Result<Vec<RouteRequest>, String> {
         let mut routes: Vec<RouteRequest> = vec![];
 
         for (path, path_item) in paths.iter() {
@@ -186,7 +198,8 @@ mod internal {
 
                     for (method, method_operation) in item.iter() {
                         let route =
-                            get_route_from_path_item(method, method_operation, &path_pattern)?;
+                            get_route_from_path_item(method, method_operation, &path_pattern, ctx)
+                                .await?;
                         routes.push(route);
                     }
                 }
@@ -202,10 +215,11 @@ mod internal {
         Ok(routes)
     }
 
-    pub(crate) fn get_route_from_path_item(
+    pub(super) async fn get_route_from_path_item(
         method: &str,
         method_operation: &Operation,
         path_pattern: &AllPathPatterns,
+        ctx: &BoxConversionContext<'_>,
     ) -> Result<RouteRequest, String> {
         let method_res = match method {
             "get" => Ok(MethodPattern::Get),
@@ -258,7 +272,7 @@ mod internal {
                     }
 
                     (GatewayBindingType::Default, _) => {
-                        let binding = get_worker_binding(worker_gateway_info)?;
+                        let binding = get_worker_binding(worker_gateway_info, ctx).await?;
 
                         Ok(RouteRequest {
                             path: path_pattern.clone(),
@@ -269,7 +283,7 @@ mod internal {
                         })
                     }
                     (GatewayBindingType::FileServer, _) => {
-                        let binding = get_worker_binding(worker_gateway_info)?;
+                        let binding = get_worker_binding(worker_gateway_info, ctx).await?;
 
                         Ok(RouteRequest {
                             path: path_pattern.clone(),
@@ -280,7 +294,7 @@ mod internal {
                         })
                     }
                     (GatewayBindingType::HttpHandler, _) => {
-                        let binding = get_http_handler_binding(worker_gateway_info)?;
+                        let binding = get_http_handler_binding(worker_gateway_info, ctx).await?;
 
                         Ok(RouteRequest {
                             path: path_pattern.clone(),
@@ -317,12 +331,21 @@ mod internal {
         }
     }
 
-    pub(crate) fn get_worker_binding(
+    pub(super) async fn get_worker_binding(
         gateway_binding_value: &Value,
+        ctx: &BoxConversionContext<'_>,
     ) -> Result<WorkerBinding, String> {
+        let component_name = get_component_name(gateway_binding_value)?;
+        let component_version = get_component_version(gateway_binding_value)?;
+
+        let component_id = ctx.resolve_component_id(&component_name).await?;
+
         let binding = WorkerBinding {
             worker_name: get_worker_id_expr(gateway_binding_value)?,
-            component_id: get_component_id(gateway_binding_value)?,
+            component_id: VersionedComponentId {
+                component_id,
+                version: component_version,
+            },
             idempotency_key: get_idempotency_key(gateway_binding_value)?,
             response_mapping: get_response_mapping(gateway_binding_value)?,
             invocation_context: get_invocation_context(gateway_binding_value)?,
@@ -331,19 +354,28 @@ mod internal {
         Ok(binding)
     }
 
-    pub(crate) fn get_http_handler_binding(
+    pub(super) async fn get_http_handler_binding(
         gateway_binding_value: &Value,
+        ctx: &BoxConversionContext<'_>,
     ) -> Result<HttpHandlerBinding, String> {
+        let component_name = get_component_name(gateway_binding_value)?;
+        let component_version = get_component_version(gateway_binding_value)?;
+
+        let component_id = ctx.resolve_component_id(&component_name).await?;
+
         let binding = HttpHandlerBinding {
             worker_name: get_worker_id_expr(gateway_binding_value)?,
-            component_id: get_component_id(gateway_binding_value)?,
+            component_id: VersionedComponentId {
+                component_id,
+                version: component_version,
+            },
             idempotency_key: get_idempotency_key(gateway_binding_value)?,
         };
 
         Ok(binding)
     }
 
-    pub(crate) fn get_cors_static_binding(
+    pub(super) fn get_cors_static_binding(
         worker_gateway_info: &Value,
     ) -> Result<StaticBinding, String> {
         match worker_gateway_info {
@@ -367,31 +399,27 @@ mod internal {
         }
     }
 
-    pub(crate) fn get_component_id(
-        gateway_binding_value: &Value,
-    ) -> Result<VersionedComponentId, String> {
-        let component_id_str = gateway_binding_value
-            .get("component-id")
-            .ok_or("No component-id found")?
+    pub(super) fn get_component_name(gateway_binding_value: &Value) -> Result<String, String> {
+        let result = gateway_binding_value
+            .get("component-name")
+            .ok_or("No component-name found")?
             .as_str()
-            .ok_or("component-id is not a string")?;
+            .ok_or("component-name is not a string")?;
 
-        let version = gateway_binding_value
+        Ok(result.to_string())
+    }
+
+    pub(super) fn get_component_version(gateway_binding_value: &Value) -> Result<u64, String> {
+        let result = gateway_binding_value
             .get("component-version")
             .ok_or("No component-version found")?
             .as_u64()
             .ok_or("component-version is not a u64")?;
 
-        let component_id =
-            ComponentId(Uuid::parse_str(component_id_str).map_err(|err| err.to_string())?);
-
-        Ok(VersionedComponentId {
-            component_id,
-            version,
-        })
+        Ok(result)
     }
 
-    pub(crate) fn get_binding_type(
+    pub(super) fn get_binding_type(
         worker_gateway_info: &Value,
     ) -> Result<GatewayBindingType, String> {
         let binding_type_optional: Option<GatewayBindingType> = worker_gateway_info
@@ -403,7 +431,7 @@ mod internal {
         Ok(binding_type_optional.unwrap_or(GatewayBindingType::Default))
     }
 
-    pub(crate) fn get_response_mapping(
+    pub(super) fn get_response_mapping(
         gateway_binding_value: &Value,
     ) -> Result<ResponseMapping, String> {
         let response = {
@@ -424,7 +452,7 @@ mod internal {
         Ok(ResponseMapping(response.clone()))
     }
 
-    pub(crate) fn get_worker_id_expr(
+    pub(super) fn get_worker_id_expr(
         gateway_binding_value: &Value,
     ) -> Result<Option<Expr>, String> {
         let worker_id_str_opt = gateway_binding_value
@@ -439,7 +467,7 @@ mod internal {
         Ok(worker_id_expr_opt)
     }
 
-    pub(crate) fn get_idempotency_key(
+    pub(super) fn get_idempotency_key(
         gateway_binding_value: &Value,
     ) -> Result<Option<Expr>, String> {
         if let Some(key) = gateway_binding_value.get("idempotency-key") {
@@ -452,7 +480,7 @@ mod internal {
         }
     }
 
-    pub(crate) fn get_invocation_context(
+    pub(super) fn get_invocation_context(
         gateway_binding_value: &Value,
     ) -> Result<Option<Expr>, String> {
         if let Some(key) = gateway_binding_value.get("invocation-context") {
@@ -465,26 +493,47 @@ mod internal {
         }
     }
 
-    pub(crate) fn get_path_pattern(path: &str) -> Result<AllPathPatterns, String> {
+    pub(super) fn get_path_pattern(path: &str) -> Result<AllPathPatterns, String> {
         AllPathPatterns::parse(path)
     }
 }
 
 #[cfg(test)]
 mod tests {
+    use golem_common::model::ComponentId;
+    use golem_service_base::model::ComponentName;
     use test_r::test;
 
     use super::*;
     use crate::gateway_api_definition::http::{AllPathPatterns, MethodPattern, RouteRequest};
     use crate::gateway_binding::{GatewayBinding, StaticBinding};
     use crate::gateway_middleware::HttpCors;
-
+    use crate::service::gateway::ConversionContext;
+    use async_trait::async_trait;
     use openapiv3::Operation;
-
     use serde_json::json;
 
+    struct DummyConversionCtx;
+
+    #[async_trait]
+    impl ConversionContext for DummyConversionCtx {
+        async fn resolve_component_id(&self, _name: &str) -> Result<ComponentId, String> {
+            unimplemented!()
+        }
+        async fn get_component_name(
+            &self,
+            _component_id: &ComponentId,
+        ) -> Result<ComponentName, String> {
+            unimplemented!()
+        }
+    }
+
+    fn make_dummy_conversion_context() -> BoxConversionContext<'static> {
+        Box::new(DummyConversionCtx)
+    }
+
     #[test]
-    fn test_get_route_with_cors_preflight_binding() {
+    async fn test_get_route_with_cors_preflight_binding() {
         let path_item = Operation {
             extensions: vec![(
                 "x-golem-api-gateway-binding".to_string(),
@@ -500,14 +549,20 @@ mod tests {
 
         let path_pattern = AllPathPatterns::parse("/test").unwrap();
 
-        let result = get_route_from_path_item("options", &path_item, &path_pattern);
+        let result = get_route_from_path_item(
+            "options",
+            &path_item,
+            &path_pattern,
+            &make_dummy_conversion_context(),
+        )
+        .await;
 
         let expected = expected_route_with_cors_preflight_binding(&path_pattern);
         assert_eq!(result, Ok(expected));
     }
 
     #[test]
-    fn test_get_route_with_cors_preflight_binding_default_response() {
+    async fn test_get_route_with_cors_preflight_binding_default_response() {
         let path_item = Operation {
             extensions: vec![(
                 "x-golem-api-gateway-binding".to_string(),
@@ -522,19 +577,31 @@ mod tests {
 
         let path_pattern = AllPathPatterns::parse("/test").unwrap();
 
-        let result = get_route_from_path_item("options", &path_item, &path_pattern);
+        let result = get_route_from_path_item(
+            "options",
+            &path_item,
+            &path_pattern,
+            &make_dummy_conversion_context(),
+        )
+        .await;
 
         let expected = expected_route_with_cors_preflight_binding_default(&path_pattern);
         assert_eq!(result, Ok(expected));
     }
 
     #[test]
-    fn test_get_route_with_no_binding_with_options_method() {
+    async fn test_get_route_with_no_binding_with_options_method() {
         let path_item = Operation::default();
 
         let path_pattern = AllPathPatterns::parse("/test").unwrap();
 
-        let result = get_route_from_path_item("options", &path_item, &path_pattern);
+        let result = get_route_from_path_item(
+            "options",
+            &path_item,
+            &path_pattern,
+            &make_dummy_conversion_context(),
+        )
+        .await;
 
         let expected = expected_route_with_cors_preflight_binding_default(&path_pattern);
         assert_eq!(result, Ok(expected));
